+++ rollup.config.js
@@ -39,4 +39,5 @@ export default {
       ],
     }),
   ],
+  external: ['react', 'react-dom', 'prop-types'],
 };