+++ src/ProSidebar/ProSidebar.tsx
@@ -1,4 +1,4 @@
-import React, { forwardRef, LegacyRef } from 'react';
+import React, { forwardRef, LegacyRef, createContext, useEffect, useState } from 'react';
 import '../scss/styles.scss';
 import classNames from 'classnames';
 
@@ -11,21 +11,44 @@ export interface Props {
   children?: React.ReactNode;
 }
 
+export interface SidebarContextProps {
+  collapsed: boolean;
+  rtl: boolean;
+}
+
+export const SidebarContext = createContext<SidebarContextProps>({
+  collapsed: false,
+  rtl: false,
+});
+
 const ProSidebar: React.ForwardRefRenderFunction<unknown, Props> = (
   { children, className, width, collapsed, rtl, image, ...rest },
   ref,
 ) => {
+  const [sidebarState, setSidebarState] = useState({ collapsed: false, rtl: false });
+
   const sidebarRef: LegacyRef<HTMLDivElement> = (ref as any) || React.createRef<HTMLDivElement>();
+
+  useEffect(() => {
+    setSidebarState({ ...sidebarState, collapsed });
+  }, [collapsed]);
+
+  useEffect(() => {
+    setSidebarState({ ...sidebarState, rtl });
+  }, [rtl]);
+
   return (
-    <div
-      {...rest}
-      ref={sidebarRef}
-      className={classNames('pro-sidebar', className, { collapsed, rtl })}
-      style={{ width }}
-    >
-      {image ? <img src={image} alt="sidebar background" className="sidebar-bg" /> : null}
-      <div className="pro-sidebar-inner">{children}</div>
-    </div>
+    <SidebarContext.Provider value={sidebarState}>
+      <div
+        {...rest}
+        ref={sidebarRef}
+        className={classNames('pro-sidebar', className, { collapsed, rtl })}
+        style={{ width }}
+      >
+        {image ? <img src={image} alt="sidebar background" className="sidebar-bg" /> : null}
+        <div className="pro-sidebar-inner">{children}</div>
+      </div>
+    </SidebarContext.Provider>
   );
 };